Use standardized official library name prefix
The names of all new official libraries must start with "Arduino_". This creates a "namespace" to avoid collisions with
names already claimed by 3rd party libraries and differentiates the official offerings from unvetted 3rd party libraries.

It's best practices for the following three things to be the same:

- Repository name
- library.properties name value
- Primary header file name

The library installation folder name is determined by the name value when installed via Library Manager.
The library installation folder name is determined by the repository name when installed via git clone or a .zip file download.
The reason it's important for the primary header file name to match the installation folder is because
[File/folder name matching](https://arduino.github.io/arduino-cli/latest/sketch-build-process/#folder-name-priority) is
one factor used to determine priority during dependency resolution.

For this reason, this commit renames the header file in parallel with the name value. It is not possible to rename the
repository via a commit so that must be done separately.