Add license file
This file is required for GitHub's [automatic license detection feature](https://docs.github.com/en/github/creating-cloning-and-archiving-repositories/licensing-a-repository).

This is the exact unmodified text of the LGPL 2.1 license from https://choosealicense.com/licenses/lgpl-2.1

Note that this text should not be modified in any way. Doing so may cause the license to no longer be recognize by GitHub
and also may affect its function as a legal instrument.