Reconfigure Arduino Lint CI for Library Manager (#24)
Arduino Lint, which is used in the "Check Arduino" CI workflow, checks for compliance with the requirements for
submission of a library into the Library Manager index, as well as for the requirements for releases of the library to be
indexed once it is accepted. Since some of these requirements are contradictory, Arduino Lint must be configured according to the library's status.

In GitHub Actions workflows, this configuration is done via the `arduino/arduino-lint-action` action's `library-manager`
input.

Now that the library has been accepted into the Library Manager index, the "update" configuration must be used.